Fix NSError in ErrorVC
