+++ MercadoPagoSDK/MercadoPagoSDK/ErrorViewController.swift
@@ -19,14 +19,14 @@ open class ErrorViewController: MercadoPagoUIViewController {
     @IBOutlet weak var exitButton: MPButton!
     
     @IBOutlet weak var retryButton: MPButton!
-    var error : MPXError!
+    var error : MPSDKError!
     var callback : ((Void) -> Void)?
     
     override open var screenName : String { get { return "ERROR" } }
     
     open var exitErrorCallback : ((Void) -> Void)!
     
-    public init(error : MPXError!, callback : ((Void) -> Void)?, callbackCancel : ((Void) -> Void)? = nil){
+    public init(error : MPSDKError!, callback : ((Void) -> Void)?, callbackCancel : ((Void) -> Void)? = nil){
         super.init(nibName: "ErrorViewController", bundle: MercadoPago.getBundle())
         self.error = error
         self.exitErrorCallback = {
