+++ MercadoPagoSDK/MercadoPagoSDK/SecrurityCodeViewController.swift
@@ -230,7 +230,7 @@ open class SecrurityCodeViewModel: NSObject {
                 self.callback(token)
                 }, failure: { (error) in
                     self.vc.hideLoading()
-                    let mpError =  MPXError(message: "Hubo un error".localized, messageDetail: error.description, retry: false)
+                    let mpError =  MPSDKError(message: "Hubo un error".localized, messageDetail: error.description, retry: false)
                     self.vc.displayFailure(mpError)
             })
         }
@@ -247,7 +247,7 @@ open class SecrurityCodeViewModel: NSObject {
             self.callback(token)
             }, failure: { (error) in
                 self.vc.hideLoading()
-                 let mpError =  MPXError(message: "Hubo un error".localized, messageDetail: error.description, retry: false)
+                 let mpError =  MPSDKError(message: "Hubo un error".localized, messageDetail: error.description, retry: false)
                 self.vc.displayFailure(mpError)
            })
